requirements.txt
# Data Manipulation & ML
pandas
scikit-learn
joblib
numpy

# Format Data & Konektor
pyarrow
boto3
minio

# Streaming
kafka-python==2.0.2
lz4

# API Service
fastapi
uvicorn[standard]

# Dashboard & Visualisasi
streamlit
plotly